Custom Designed Reinforcement Learning Project
This is a Machine Learning model I created inspired by the NEAT algorithm. Like the mentioned algorithm, this one uses the concept of Neuro Evolution to design a network structure (Topology) best fit for solving a particular problem. However unlike NEAT, this algorithm does not use speciation, and instead uses copies of older parents to ensure stable learning, and, in addition is able to remove particular Genes through the process of "Hanging Connections"